+++ src/modules/preprocessing.py
@@ -8,10 +8,7 @@ import textblob as tb
 import yaml
 from nltk.corpus import stopwords as sw
 from nltk.stem import PorterStemmer, WordNetLemmatizer
-from numpy.typing import ArrayLike
-from pandas import DataFrame, Series
-from rapidfuzz.fuzz import ratio
-from sklearn.feature_extraction.text import CountVectorizer
+from pandas import Series
 
 
 def load_config(filepath: str) -> dict:
@@ -125,24 +122,6 @@ def _update_spelling_words(additional_words: dict) -> None:
     return tb.en.spelling
 
 
-def fuzzy_compare_ratio(base: Series, comparison: Series) -> Series:
-    """compare the base series to the comparison series to get
-    a similarity ratio between strings in the same column
-    Parameters
-    ----------
-    base: Series
-        the base series for comparison
-    comparison: Series
-        the series you want to compare against
-    Returns
-    -------
-    Series
-        a series of ratios (type:float) with scores closer to 100
-        indicating complete match"""
-    fuzzy_ratio = Series(map(ratio, base, comparison))
-    return fuzzy_ratio
-
-
 def remove_punctuation(text: str) -> str:
     """Remove punctuation from string
 
@@ -156,6 +135,7 @@ def remove_punctuation(text: str) -> str:
     str
         text string without punctuation
     """
+    _initialise_nltk_component("tokenizers/punkt", "punkt")
     new_text = re.sub(string=text, pattern="[{}]".format(string.punctuation), repl="")
     return new_text
 
@@ -207,9 +187,30 @@ def lemmatizer(tokens: list) -> list:
     lemmatized_tokens
         list of simplified word groupings
     """
+    _initialise_nltk_component("corpora/wordnet.zip", "wordnet")
     lemmatizer = WordNetLemmatizer()
     lemmatized_tokens = [lemmatizer.lemmatize(token) for token in tokens]
     return lemmatized_tokens
+
+
+def _initialise_nltk_component(extension: str, download_object: str):
+    """download nltk component from package
+    Parameters
+    ----------
+    extension: str
+        the filepath extension leading to where the model is saved
+    download_object: str
+        the object to download from nltk
+    Returns
+    -------
+    None
+    """
+    username = os.getenv("username")
+    path = "c:/Users/" + username + "/AppData/Roaming/nltk_data" + extension
+    if not os.path.exists(path):
+        nltk.download(download_object)
+    nltk.data.path.append("../local_packages/nltk_data")
+    return None
 
 
 def remove_nltk_stopwords(tokens: list, additional_stopwords: list = []) -> list:
@@ -243,26 +244,10 @@ def initialise_update_stopwords(additional_stopwords: list = None) -> list:
     list
         a list of words to remove from corpus
     """
-    stopwords = _initialise_nltk_stopwords()
+    _initialise_nltk_component("corpora/stopwords", "stopwords")
+    stopwords = sw.words("english")
     updated_stopwords = _update_nltk_stopwords(stopwords, additional_stopwords)
     return updated_stopwords
-
-
-def _initialise_nltk_stopwords() -> list:
-    """fetch nltk stopwords from corpora
-
-    Returns
-    -------
-    list
-        list of nltk stopwords
-    """
-    username = os.getenv("username")
-    path = "c:/Users/" + username + "/AppData/Roaming/nltk_data/corpora/stopwords"
-    if not os.path.exists(path):
-        nltk.download("stopwords")
-    nltk.data.path.append("../local_packages/nltk_data")
-    stopwords = sw.words("english")
-    return stopwords
 
 
 def _update_nltk_stopwords(stopwords: list, additional_stopwords: list):
@@ -298,81 +283,3 @@ def rejoin_tokens(tokens: list) -> str:
     """
     joined_tokens = " ".join(tokens)
     return joined_tokens
-
-
-def extract_feature_count(
-    series: Series,
-    max_features: int = None,
-    ngram_range: tuple[float, float] = (1, 1),
-    stop_words: ArrayLike = None,
-    lowercase: bool = True,
-    min_df=1,
-    max_df=1.0,
-):
-    """create a text feature count dataframe from series
-    Paramaters
-    ----------
-    series: Series
-        Series of text strings
-    max_features: int, default = None
-        If not None, build a vocabulary that only consider the top max_features
-        ordered by term frequency across the corpus. Otherwise, all features are used.
-    ngram_range: tuple (min_n, max_n), default=(1, 1)
-        The lower and upper boundary of the range of n-values for different word n-grams
-        or char n-grams to be extracted. All values of n such such that
-        min_n <= n <= max_n will be used.
-    stop_words: list, default=None
-        list of stopwords to remove from text strings
-    lowercase: bool, default = True
-        convert all characters to lowercase before tokenizing
-    min_df: float or int, default = 1
-        When building the vocabulary ignore terms that have a document frequency
-        strictly lower than the given threshold. This value is also called cut-off
-        in the literature. If float, the parameter represents a proportion of
-        documents, integer absolute counts.
-        This parameter is ignored if vocabulary is not None.
-    max_df: float or int, default = 1.0
-        When building the vocabulary ignore terms that have a document frequency
-        strictly higher than the given threshold (corpus-specific stop words).
-        If float, the parameter represents a proportion of documents, integer
-        absolute counts. This parameter is ignored if vocabulary is not None.
-    Returns
-    -------
-    DataFrame
-        A dataframe of text feature counts, displaying the number of times a word
-        appears in each element of the input series
-    """
-
-    vectorizer = CountVectorizer(
-        max_features=max_features,
-        ngram_range=ngram_range,
-        stop_words=stop_words,
-        lowercase=lowercase,
-        min_df=min_df,
-        max_df=max_df,
-    )
-
-    fitted_vector = vectorizer.fit_transform(series)
-
-    word_count_df = DataFrame(
-        fitted_vector.toarray(), columns=vectorizer.get_feature_names_out()
-    )
-    return word_count_df
-
-
-def get_total_feature_count(features: DataFrame) -> DataFrame:
-    """sum across features to get total number of times word was used
-    Parameters
-    ----------
-    features: DataFrame
-        A dataframe of the features with each row corrosponding to a deconstructed
-        string
-    Returns
-    -------
-    DataFrame
-        A dataframe of the total number of times each word is used across all
-        strings"""
-    total_feature_count = DataFrame()
-    for column in features.columns:
-        total_feature_count[column] = [features[column].sum()]
-    return total_feature_count
